requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.3
brotli==1.2.0
celery==5.5.3
certifi==2025.10.5
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
geographiclib==2.1
geopy==2.4.1
gunicorn==23.0.0
h11==0.16.0
idna==3.11
kombu==5.5.4
msgpack==1.1.2
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
razorpay==2.0.0
redis==7.0.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
